Changes for SUT files# Please enter the commit message for your changes. Lines starting

+++ bin/start_client.py
@@ -650,20 +650,28 @@ class TwisterClientService(rpyc.Service):
         """ save sut to file """
 
         # Save sut files
+        print('\nSave SUTS {}\n'.format(sutList))
         for (name, sut) in sutList:
             try:
+                print('\n BOG Name {} Sut {}\n'.format(name, sut))
                 sutsPath = self._conn.root.getUserVariable('sut_path')
                 if not sutsPath:
                     sutsPath = '{}/config/sut/'.format(TWISTER_PATH)
                 childPath = os.path.join(sutsPath, name)
-                with open(childPath, 'r') as f:
-                    sut = json.loads(json.dumps(copy.deepcopy(sut)))
-                    f_content = json.load(f)
-                    diff = set(o for o in set(sut.keys()).intersection(f_content) if f_content[o] != sut[o])
-                    if diff:
-                        with open(childPath, 'w') as _f:
-                            json.dump(sut, _f, indent=4)
+                print('BOG: Child path {}'.format(childPath))
+                if os.path.isfile(childPath):
+                    with open(childPath, 'r') as f:
+                        sut = json.loads(json.dumps(copy.deepcopy(sut)))
+                        f_content = json.load(f)
+                        diff = set(o for o in set(sut.keys()).intersection(f_content) if f_content[o] != sut[o])
+                        if diff:
+                            with open(childPath, 'w') as _f:
+                                json.dump(sut, _f, indent=4)
+                else:
+                     with open(childPath, 'w') as _f:
+                         json.dump(sut, _f, indent=4)
             except Exception as e:
+                print('BOG: Exceptie {}'.format(e))
                 return e
 
         return True
